Car Controller Commit 108AM 11/24/2020

+++ Assets/CarController.cs
@@ -6,22 +6,22 @@ public class CarController : MonoBehaviour
 {
     bool clutchin;
 
-    int gear = 0;
-    int lastgear = 0;
-    int gearratio = 30;
-    int downgearratio = 15;
+    float gear = 0f;
+    float lastgear = 0f;
+    float gearratio = 30f;
+    float downgearratio = 15f;
 
-    int rpm = 0;
-    int maxrpm = 8500;
-    int minrpm = 0;
-    int rpmdrop = 1500;
-    int torque = 450;
+    float rpm = 0f;
+    float maxrpm = 8500f;
+    float minrpm = 0f;
+    float rpmdrop = 1500f;
+    float torque = 450f;
 
-    float inspeed = 0;
-    float speed = 0;
-    float acceleration = 0;
-    int drag = -1;
-    int hspr = 2000;
+    float inspeed = 0f;
+    float speed = 0f;
+    float acceleration = 0f;
+    float drag = -1f;
+    float hspr = 2000f;
    
     // Start is called before the first frame update
     void Start()
@@ -44,11 +44,11 @@ public class CarController : MonoBehaviour
         //sync lastgear and gear so cannot drop rpm at 6 nor rise rpm at -1
         if (gear == lastgear)
         {
-            rpmdrop = 0;
+            rpmdrop = 0f;
         }
         else if (gear != lastgear)
         {
-            rpmdrop = 1500;
+            rpmdrop = 1500f;
         }
 
         //switch gear system
@@ -57,32 +57,32 @@ public class CarController : MonoBehaviour
             if (Input.GetKeyDown(KeyCode.DownArrow))
             {
                 //so the sync system would work
-                lastgear = gear-2;
-                if (lastgear <= -1)
+                lastgear = gear-2f;
+                if (lastgear <= -1f)
                 {
-                    lastgear = -1;
+                    lastgear = -1f;
                 }
-                gear -= 1;
+                gear -= 1f;
                 //downshift rpm rise
-                if (gear != 0 && gear != -1)
+                if (gear != 0f && gear != -1f)
                 {
-                    if (rpm != 0)
+                    if (rpm != 0f)
                     {
                             rpm += rpmdrop;
                     }
                 }
                 //setting lowest gear, can only down shift back to -1 when rpm is 0
-                if (rpm != 0)
+                if (rpm != 0f)
                 {
-                    if (gear == -1)
+                    if (gear == -1f)
                     {
-                        gear = 0;
+                        gear = 0f;
                     }
-                } else if (rpm == 0)
+                } else if (rpm == 0f)
                 {
-                    if (gear <= -1)
+                    if (gear <= -1f)
                     {
-                        gear = -1;
+                        gear = -1f;
                     }
                 }
 
@@ -125,29 +125,29 @@ public class CarController : MonoBehaviour
         }
        
         //setting up lowest and highest gear
-        if (gear <= -1)
+        if (gear <= -1f)
         {
-            gear = -1;
+            gear = -1f;
         }
-        if (gear >= 6)
+        if (gear >= 6f)
         {
-            gear = 6;
+            gear = 6f;
         }
 
         //press W to increase rpm
         if (Input.GetKey(KeyCode.W))
         {
-            if (gear == 1 || gear == 2 || gear == 3 || gear == 4 || gear == 5 || gear == 6)
+            if (gear == 1f || gear == 2f || gear == 3f || gear == 4f || gear == 5f || gear == 6f)
             {
                 rpm += (torque / (gear*gearratio));
             }
 
-            if (gear == -1)
+            if (gear == -1f)
             {
-                rpm += (torque / (2 * gearratio));
+                rpm += (torque / (2f * gearratio));
             }
 
-            if (gear == 0)
+            if (gear == 0f)
             {
                 rpm += (torque / (gearratio));
             }
@@ -156,17 +156,17 @@ public class CarController : MonoBehaviour
         //release W to drop rpm
         if (Input.GetKey(KeyCode.W)==false)
         {
-            if (gear == 1 || gear == 2 || gear == 3 || gear == 4 || gear == 5 || gear == 6)
+            if (gear == 1f || gear == 2f || gear == 3f || gear == 4f || gear == 5f || gear == 6f)
             {
                 rpm -= (torque / (gear*downgearratio));
             }
 
-            if (gear == -1)
+            if (gear == -1f)
             {
                 rpm -= (torque / (2*downgearratio));
             }
 
-            if (gear == 0)
+            if (gear == 0f)
             {
                 rpm -= (torque / downgearratio);
             }
@@ -185,17 +185,17 @@ public class CarController : MonoBehaviour
 
         //setting up acceleration
         //positive acceleration when gear > 0 and rpm > 1000
-        if (gear > 0 && rpm > 1000)
+        if (gear > 0f && rpm > 1000f)
         {
             acceleration = hspr / rpm;
         }
         //no acc when gear = 0
-        else if (gear == 0)
+        else if (gear == 0f)
         {
-            acceleration = 0;
+            acceleration = 0f;
         }
         //negative acc when gear = -1 and rpm> 0
-        else if (gear == -1 && rpm > 1000)
+        else if (gear == -1f && rpm > 1000f)
         {
             acceleration = -(hspr / rpm);
         }
@@ -209,10 +209,10 @@ public class CarController : MonoBehaviour
        
 
        // print(((bool)clutchin));
-        print(((int)gear));
+        print(((float)gear));
        // print(((int)lastgear));
        // print(((int)rpmdrop));
-        print(((int)rpm));
+        print(((float)rpm));
         print(((float)acceleration));
        // print(((int)speed));
